Add code for downloading PDFs from a website using a web scraper
