+++ crawler/decover_spider.py
@@ -1,25 +1,6 @@
 import scrapy
-from bs4 import BeautifulSoup
 
-
-# TODO(Ravi): Move this to a separate file.
-def get_text_from_html(html_content):
-    soup = BeautifulSoup(html_content, features="html.parser")
-
-    # kill all script and style elements.
-    for script in soup(["script", "style"]):
-        script.extract()  # rip it out
-
-    # get text
-    text = soup.get_text()
-
-    # break into lines and remove leading and trailing space on each
-    lines = (line.strip() for line in text.splitlines())
-    # break multi-headlines into a line each
-    chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
-    # drop blank lines
-    text = ' '.join(chunk for chunk in chunks if chunk)
-    return text
+from crawler.utils.helper_methods import get_text_from_html, get_pdf_links, download_pdf
 
 
 class DecoverSpider(scrapy.Spider):
@@ -38,6 +19,9 @@ class DecoverSpider(scrapy.Spider):
 
         # Step II: Create a dictionary to store the results for this page. It will be a key-value pair of {url: text}.
         text = get_text_from_html(text_html)
+        pdf_links = get_pdf_links(text_html)
+        for pdf_link in pdf_links:
+            download_pdf(pdf_link['href'])
         result = {response.url: text}
         self.page_limit -= 1
